refactor: MemberElement 통일
