+++ frontend/src/components/organisms/MemberElement/styles.ts
